+++ src/update.ts
@@ -1,16 +1,8 @@
-import { execSync } from 'child_process'
-import { join } from 'path'
 import { PackageInstallation, removeInstallations } from './installations'
 
 import { readLockfile } from './lockfile'
 
-import {
-  parsePackageName,
-  addPackages,
-  readPackageManifest,
-  values,
-  runOrWarnPackageManagerInstall
-} from '.'
+import { parsePackageName, addPackages } from '.'
 
 export interface UpdatePackagesOptions {
   workingDir: string
@@ -28,7 +20,7 @@ export const updatePackages = async (
   let packagesToUpdate: string[] = []
   let installationsToRemove: PackageInstallation[] = []
   if (packages.length) {
-    packages.forEach(packageName => {
+    packages.forEach((packageName) => {
       const { name, version } = parsePackageName(packageName)
       if (lockfile.packages[name]) {
         if (version) {
@@ -37,7 +29,7 @@ export const updatePackages = async (
         packagesToUpdate.push(name)
       } else {
         installationsToRemove.push({ name, path: options.workingDir })
-        console.log(
+        console.warn(
           `Did not find package ${name} in lockfile, ` +
             `please use 'add' command to add it explicitly.`
         )
@@ -47,65 +39,51 @@ export const updatePackages = async (
     packagesToUpdate = Object.keys(lockfile.packages)
   }
 
-  const lockPackages = packagesToUpdate.map(name => ({
+  const lockPackages = packagesToUpdate.map((name) => ({
     name: lockfile.packages[name].version
       ? name + '@' + lockfile.packages[name].version
       : name,
     file: lockfile.packages[name].file,
     link: lockfile.packages[name].link,
-    pure: lockfile.packages[name].pure
+    pure: lockfile.packages[name].pure,
   }))
 
-  const packagesFiles = lockPackages.filter(p => p.file).map(p => p.name)
+  const packagesFiles = lockPackages.filter((p) => p.file).map((p) => p.name)
 
   const addOpts = {
     workingDir: options.workingDir,
-    replace: options.replace
+    replace: options.replace,
   }
 
   await addPackages(packagesFiles, {
     ...addOpts,
-    yarn: false
+    yarn: false,
   })
 
   const packagesLinks = lockPackages
-    .filter(p => !p.file && !p.link && !p.pure)
-    .map(p => p.name)
+    .filter((p) => !p.file && !p.link && !p.pure)
+    .map((p) => p.name)
   await addPackages(packagesLinks, {
     ...addOpts,
     link: true,
     pure: false,
-    yarn: false
+    yarn: false,
   })
 
-  const packagesLinkDep = lockPackages.filter(p => p.link).map(p => p.name)
+  const packagesLinkDep = lockPackages.filter((p) => p.link).map((p) => p.name)
   await addPackages(packagesLinkDep, {
     ...addOpts,
     linkDep: true,
     pure: false,
-    yarn: false
+    yarn: false,
   })
 
-  const packagesPure = lockPackages.filter(p => p.pure).map(p => p.name)
+  const packagesPure = lockPackages.filter((p) => p.pure).map((p) => p.name)
   await addPackages(packagesPure, {
     ...addOpts,
     pure: true,
-    yarn: false
+    yarn: false,
   })
-
-  for (const packageName of packages) {
-    const pkg = readPackageManifest(
-      join(options.workingDir, values.yalcPackagesFolder, packageName)
-    )
-    const postupdate = pkg && pkg.scripts && pkg.scripts.postupdate
-    if (postupdate) {
-      console.log(
-        `Running postupdate script of package ${packageName} in ${workingDir}`
-      )
-      execSync(postupdate, { cwd: workingDir })
-    }
-  }
-
   if (!options.noInstallationsRemove) {
     await removeInstallations(installationsToRemove)
   }